+++ src/app/api-server/src/config.rs
@@ -9,15 +9,11 @@
 
 use std::collections::BTreeMap;
 use std::path::PathBuf;
+use std::sync::Arc;
 
 use container_runtime::{ContainerRuntimeType, NetworkNamespaceType};
 use duration_string::DurationString;
 use internal_error::*;
-use kamu::{
-    EngineConfigDatafusionEmbeddedBatchQuery,
-    EngineConfigDatafusionEmbeddedCompaction,
-    EngineConfigDatafusionEmbeddedIngest,
-};
 use kamu_accounts::{AccountConfig, DidSecretEncryptionConfig};
 use kamu_accounts_services::PasswordPolicyConfig;
 use kamu_datasets::DatasetEnvVarsConfig;
@@ -183,6 +179,13 @@ pub struct EngineConfigDatafution {
 
     /// Compaction-specific overrides to the base config
     pub compaction: BTreeMap<String, String>,
+
+    /// Makes arrow batches use contiguous `Binary` and `Utf8` encodings instead
+    /// of more modern `BinaryView` and `Utf8View`. This is only needed for
+    /// compatibility with some older libraries that don't yet support them.
+    ///
+    /// See: [kamu-node#277](https://github.com/kamu-data/kamu-node/issues/277)
+    pub use_legacy_arrow_buffer_encoding: bool,
 }
 
 impl Default for EngineConfigDatafution {
@@ -204,6 +207,7 @@ impl Default for EngineConfigDatafution {
                 .iter()
                 .map(|(k, v)| ((*k).to_string(), (*v).to_string()))
                 .collect(),
+            use_legacy_arrow_buffer_encoding: false,
         }
     }
 }
@@ -235,17 +239,23 @@ impl EngineConfigDatafution {
             };
 
         let ingest_config = from_merged_with_base_and_defaults(
-            EngineConfigDatafusionEmbeddedIngest::DEFAULT_OVERRIDES,
+            kamu::EngineConfigDatafusionEmbeddedIngest::DEFAULT_OVERRIDES,
             self.ingest,
         )?;
-        let batch_query_config = from_merged_with_base_and_defaults(
-            EngineConfigDatafusionEmbeddedBatchQuery::DEFAULT_OVERRIDES,
+        let mut batch_query_config = from_merged_with_base_and_defaults(
+            kamu::EngineConfigDatafusionEmbeddedBatchQuery::DEFAULT_OVERRIDES,
             self.batch_query,
         )?;
         let compaction_config = from_merged_with_base_and_defaults(
-            EngineConfigDatafusionEmbeddedCompaction::DEFAULT_OVERRIDES,
+            kamu::EngineConfigDatafusionEmbeddedCompaction::DEFAULT_OVERRIDES,
             self.compaction,
         )?;
+
+        batch_query_config.set_extension(Arc::new(
+            kamu::EngineConfigDatafusionEmbeddedBatchQueryExt {
+                use_legacy_arrow_buffer_encoding: self.use_legacy_arrow_buffer_encoding,
+            },
+        ));
 
         Ok((
             kamu::EngineConfigDatafusionEmbeddedIngest(ingest_config),